fix(blocks): Separate blocks namespace and dictionary objects
Use Blockly.blocks.* for blocks modules, leaving the Blockly.Blocks
name for the block dictionary object.

This solves a problem with advanced compilation, where (the minified
version of Blockly.Blocks) is overwritten, replacing the dictionary
object with an empty namespace object.  This results in some block
definitions being created in the dictionary and some on the namespace
object.